test(libsecret): split libsecret by behaviour family
Split libsecret.test.js (630 LOC) into libsecret-generators.test.js
(hash/uuid/secret/base64), libsecret-jwt.test.js (generateJWT +
Supabase minting) and libsecret-env.test.js (parseDuration +
updateEnvFile/readEnvFile/getOrGenerateSecret). Shared
makeRuntime/TEST_ENV_PATH lifted into libsecret-helpers.js. Assertions
unchanged; 64 tests preserved.

Spec 0640 § C (large files), design Decision 5.
